If running on node, prefixer is an identity function
